requirements.txt
appdirs==1.4.3
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.5.4
drfdocs==0.0.11
gunicorn==19.7.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
